Updated proudct fields: protein, carbs, fats to doubles intstead of integers for better accuracy

+++ backend/src/main/java/backend/dto/CreateProductRequest.java
@@ -15,9 +15,9 @@ public class CreateProductRequest {
     private String imageUrl;
 
     private Integer calories;
-    private Integer proteinGrams;
-    private Integer carbGrams;
-    private Integer fatGrams;
+    private Double proteinGrams;
+    private Double carbGrams;
+    private Double fatGrams;
 
     private String sourceName;
     private String sourceUrl;
@@ -74,27 +74,27 @@ public class CreateProductRequest {
         this.calories = calories;
     }
 
-    public Integer getProteinGrams() {
+    public Double getProteinGrams() {
         return proteinGrams;
     }
 
-    public void setProteinGrams(Integer proteinGrams) {
+    public void setProteinGrams(Double proteinGrams) {
         this.proteinGrams = proteinGrams;
     }
 
-    public Integer getCarbGrams() {
+    public Double getCarbGrams() {
         return carbGrams;
     }
 
-    public void setCarbGrams(Integer carbGrams) {
+    public void setCarbGrams(Double carbGrams) {
         this.carbGrams = carbGrams;
     }
 
-    public Integer getFatGrams() {
+    public Double getFatGrams() {
         return fatGrams;
     }
 
-    public void setFatGrams(Integer fatGrams) {
+    public void setFatGrams(Double fatGrams) {
         this.fatGrams = fatGrams;
     }
